scripts/extract-images/requirements.txt
PyMuPDF>=1.24.0
Pillow>=10.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
